factoring out the functions into reusable utils.py
